transcript-video-worker/requirements.txt
pika==1.3.2
minio==7.2.9
PyYAML==6.0.2
backoff==2.2.1
pydantic==2.8.2
assemblyai==0.48.1
psycopg2-binary==2.9.9
requests==2.31.0

